requirements.txt
colorama==0.4.6
dnspython==2.7.0
iniconfig==2.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pluggy==1.6.0
pymongo==4.13.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
